requirements.txt
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.2
python-dotenv==1.2.1
